Minimize number of pings in the network status checks
Add an implied dependency among the network status tests so we can
assume a test is ok if the following test succeed. This minimizes the
number of pings we are required to send, but also require the test
list to have a natural dependency order (e.g. VPN must depend on DNS
to work or we may get a false positive).