SMS handler updates
Updates to SMS handler to prevent responding to bots or carrier messages (need a more elegant way of doing so other than just matching the prefix). Also tweaked wording of some responders. I am trying to be as short as possible to fit a single message, but still able to explain functions.