Changed default destination to APZIOR
Change default destination to APZIOR as an identifier. The "Z" stands for experimental software. Added "IOR" to identify it as ioreth. Might be useful to differentiate it from other applications that use the default APRS destination, in case igate managers or other users want to do some filtering based on dest.